+++ Image microservice/.gitignore
@@ -1,3 +1,4 @@
 # dependencies
 /node_modules
 /dist
+.env
