+++ Image microservice/jest.config.js
