+++ Image microservice/package.json
@@ -3,6 +3,7 @@
     "amqplib": "^0.10.3",
     "cors": "^2.8.5",
     "dayjs": "^1.11.6",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.0",
     "express-jwt": "^7.7.5",
@@ -17,16 +18,21 @@
     "@types/body-parser": "^1.19.2",
     "@types/cors": "^2.8.12",
     "@types/express-fileupload": "^1.4.1",
+    "@types/jest": "^29.2.5",
     "@types/morgan": "^1.9.3",
+    "@types/supertest": "^2.0.12",
+    "jest": "^29.3.1",
     "nodemon": "^2.0.19",
     "pino-pretty": "^9.1.1",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.3",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.8.3"
   },
   "scripts": {
-    "dev": "nodemon app.ts",
+    "dev": "nodemon server.ts",
     "build": "tsc",
-    "start": "tsc && node dist/app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "tsc && node dist/server.js",
+    "test": "jest"
   }
 }
