+++ Image microservice/src/app.ts
@@ -1,12 +1,8 @@
 import express from 'express';
 import cors from 'cors';
-import log from './src/utils/logger';
 import morgan from 'morgan';
-import ImageRoutes from './src/routes/imageRoutes';
 import fileUpload from 'express-fileupload';
-import { onProductUpdated } from './src/utils/rabbitmq';
 
-const port = 3010;
 const app = express();
 
 app.use(cors({ origin: 'http://localhost:3000' }));
@@ -19,8 +15,4 @@ app.use(
   })
 );
 
-app.listen(port, async () => {
-  log.info(`App is running at http://localhost:${port}`);
-  onProductUpdated();
-  ImageRoutes(app);
-});
+export default app;
