+++ Product microservice/.gitignore
@@ -2,3 +2,5 @@
 /node_modules
 /dist
 /public/**
+.env
+coverage
