+++ Product microservice/package.json
@@ -2,14 +2,14 @@
   "name": "s-a-rb03-individual",
   "version": "1.0.0",
   "description": "To make it easy for you to get started with GitLab, here's a list of recommended next steps.",
-  "main": "dist/app.js",
+  "main": "dist/server.js",
   "scripts": {
     "predev": "npm run swagger",
     "prebuild": "npm run swagger",
     "dev": "concurrently \"nodemon\" \"nodemon -x tsoa spec\"",
     "build": "tsc",
-    "start": "tsc && node dist/app.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "tsc && node dist/server.js",
+    "test": "jest",
     "swagger": "tsoa spec"
   },
   "repository": {
@@ -48,14 +48,16 @@
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/express-fileupload": "^1.4.1",
-    
+    "@types/jest": "^29.2.5",
     "@types/morgan": "^1.9.3",
     "@types/node": "^18.7.16",
+    "@types/supertest": "^2.0.12",
     "@types/swagger-ui-express": "^4.1.3",
     "concurrently": "^7.5.0",
-    
+    "jest": "^29.3.1",
     "nodemon": "^2.0.19",
-
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.3",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.8.3"
@@ -65,6 +67,6 @@
       "./"
     ],
     "ext": "ts",
-    "exec": "ts-node app.ts"
+    "exec": "ts-node server.ts"
   }
 }
