+++ Product microservice/tsoa.json
@@ -1,5 +1,5 @@
 {
-    "entryFile": "app.ts",
+    "entryFile": "server.ts",
     "noImplicitAdditionalProperties": "throw-on-extras",
     "spec": {
       "outputDirectory": "public",
