+++ Shopping cart microservice/.dockerignore
@@ -1,1 +1,2 @@
 node_modules
+test
