+++ Shopping cart microservice/.gitignore
@@ -1,4 +1,4 @@
 # dependencies
 /node_modules
 /dist
-
+.env
