--- Shopping cart microservice/app.ts
@@ -1,21 +1,0 @@
-import express from 'express';
-import config from 'config';
-import connect from './src/utils/db.connect';
-import logger from './src/utils/logger';
-import { onProductDeleted, onProductUpdated } from './src/utils/rabbitmq';
-require('dotenv').config();
-import shoppingCartRoutes from './src/routes/shoppingCartRoutes';
-
-const app = express();
-
-const port = process.env.PORT || 3009;
-
-app.use(express.json());
-
-app.listen(port, async () => {
-  logger.info(`App is running at http://localhost:${port}`);
-  await connect();
-  onProductUpdated();
-  onProductDeleted();
-  shoppingCartRoutes(app);
-});
