+++ Shopping cart microservice/package.json
@@ -4,10 +4,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon app.ts",
+    "dev": "nodemon server.ts",
     "build": "tsc",
-    "start": "tsc && node dist/app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "tsc && node dist/server.js",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
@@ -30,12 +30,19 @@
     "js-guid": "^1.0.0",
     "jwks-rsa": "^2.1.5",
     "mongoose": "^6.6.0",
+    "morgan": "^1.10.0",
     "pino": "^8.5.0",
     "zod": "^3.19.1"
   },
   "devDependencies": {
+    "@types/jest": "^29.2.5",
+    "@types/morgan": "^1.9.4",
+    "@types/supertest": "^2.0.12",
+    "jest": "^29.3.1",
     "nodemon": "^2.0.19",
     "pino-pretty": "^9.1.1",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.0.3",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.8.3"
   }
