+++ Shopping cart microservice/src/app.ts
@@ -1,0 +1,9 @@
+import express from 'express';
+import morgan from 'morgan';
+
+const app = express();
+
+app.use(express.json());
+app.use(morgan('dev'));
+
+export default app;
